Split body management into its own module
